+++ core/src/main/java/com/asml/apa/wta/core/supplier/SupplierExtractionEngine.java
@@ -3,6 +3,7 @@ package com.asml.apa.wta.core.supplier;
 import com.asml.apa.wta.core.dto.BaseSupplierDto;
 import com.asml.apa.wta.core.dto.IostatDto;
 import com.asml.apa.wta.core.dto.OsInfoDto;
+import com.asml.apa.wta.core.dto.PerfDto;
 import com.asml.apa.wta.core.utils.BashUtils;
 import java.time.LocalDateTime;
 import java.util.ArrayList;
@@ -26,6 +27,8 @@ public abstract class SupplierExtractionEngine<T extends BaseSupplierDto> {
 
   private final IostatSupplier iostatSupplier;
 
+  private final PerfSupplier perfSupplier;
+
   private final int resourcePingInterval;
 
   @Getter
@@ -45,6 +48,7 @@ public abstract class SupplierExtractionEngine<T extends BaseSupplierDto> {
     this.resourcePingInterval = resourcePingInterval;
     this.operatingSystemSupplier = new OperatingSystemSupplier();
     this.iostatSupplier = new IostatSupplier(new BashUtils());
+    this.perfSupplier = new PerfSupplier(new BashUtils());
   }
 
   /**
@@ -53,19 +57,22 @@ public abstract class SupplierExtractionEngine<T extends BaseSupplierDto> {
    *
    * @return A {@link CompletableFuture} that completes when the result has been resolved
    * @author Henry Page
+   * @author Pil Kyu Cho
    * @since 1.0.0
    */
   protected CompletableFuture<T> ping() {
     CompletableFuture<OsInfoDto> osInfoDtoCompletableFuture = this.operatingSystemSupplier.getSnapshot();
     CompletableFuture<IostatDto> iostatDtoCompletableFuture = this.iostatSupplier.getSnapshot();
+    CompletableFuture<PerfDto> perfDtoCompletableFuture = this.perfSupplier.getSnapshot();
 
     return CompletableFuture.allOf(osInfoDtoCompletableFuture, iostatDtoCompletableFuture)
         .thenCompose((v) -> {
           LocalDateTime timestamp = LocalDateTime.now();
           OsInfoDto osInfoDto = osInfoDtoCompletableFuture.join();
           IostatDto iostatDto = iostatDtoCompletableFuture.join();
+          PerfDto perfDto = perfDtoCompletableFuture.join();
           return CompletableFuture.completedFuture(
-              transform(new BaseSupplierDto(timestamp, osInfoDto, iostatDto)));
+              transform(new BaseSupplierDto(timestamp, osInfoDto, iostatDto, perfDto)));
         });
   }