+++ CHANGELOG.md
@@ -53,6 +53,7 @@ For the most part, since we are pre-release this will serve no real purpose, oth
 - Modified the config format [@amousavigourab](https://gitlab.ewi.tudelft.nl/amousavigourab)
 - Modified the EndToEnd class to inject the WtaPlugin class into a Spark context rather the Spark listener interface by itself [@pcho](https://gitlab.ewi.tudelft.nl/pcho)
 - Reworked the Docker images from the ground up to use Ubuntu Focal [@amousavigourab](https://gitlab.ewi.tudelft.nl/amousavigourab)
+- Refactor energy consumption metrics data source to be synchronized with driver [@pcho](https://gitlab.ewi.tudelft.nl/pcho)
 
 ### Fixed
 - Fixed the slf4j logging module to allow use in testing and `adapter/spark` [@amousavigourab](https://gitlab.ewi.tudelft.nl/amousavigourab)